requirements.txt
tensorflow>=2.11
numpy
matplotlib
pillow
gradio
